cron: email warnings after 18 hours and queue reverting after 24 hours.
Note:
A. This isn't turned on yet. The CRON_RUNNING setting changes this.
B. It works once an hour, so not exact. I don't think it needs to be.
C. There are a couple of TODOs, e.g. one to change to using aggregate
D. The reverting isn't yet robust to editors just re-claiming after it runs out.
   I am aware of this. It would need another field in episode to be aware that this had been reverted before. We could also just not care for now as it's not imperative to fix.
   If the person really wants to do it again, I'm k with letting them.